разбивает все на модули
